+++ src/main/java/com/pht/rinha_backend_2026/dto/FraudResponse.java
@@ -1,4 +1,4 @@
 package com.pht.rinha_backend_2026.dto;
 
-public record FraudResponse(boolean approved, double fraud_score) {
+public record FraudResponse(boolean approved, float fraud_score) {
 }
