+++ src/main/java/com/pht/rinha_backend_2026/dto/LastTransaction.java
@@ -2,5 +2,5 @@ package com.pht.rinha_backend_2026.dto;
 
 import java.time.Instant;
 
-public record LastTransaction(Instant timestamp, double km_from_current) {
+public record LastTransaction(Instant timestamp, float km_from_current) {
 }
